requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyJWT[crypto]==2.8.0
python-dotenv==1.0.1
requests==2.32.3
